[yarpGPT] Change namespace. Test case correction
