+++ src/devices/ILLMMsgs/ILLMMsgs.thrift
@@ -3,6 +3,8 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
+namespace yarp yarp.dev.llm
+
 struct Message {
     string sender;
     string content;
