+++ src/devices/ILLMMsgs/idl_generated_code/ILLMMsgs_index.txt
@@ -1,10 +1,10 @@
-Message.h
-Message.cpp
-return_readPrompt.h
-return_readPrompt.cpp
-return_ask.h
-return_ask.cpp
-return_getConversation.h
-return_getConversation.cpp
-ILLMMsgs.h
-ILLMMsgs.cpp
+yarp/dev/llm/Message.h
+yarp/dev/llm/Message.cpp
+yarp/dev/llm/return_readPrompt.h
+yarp/dev/llm/return_readPrompt.cpp
+yarp/dev/llm/return_ask.h
+yarp/dev/llm/return_ask.cpp
+yarp/dev/llm/return_getConversation.h
+yarp/dev/llm/return_getConversation.cpp
+yarp/dev/llm/ILLMMsgs.h
+yarp/dev/llm/ILLMMsgs.cpp
