+++ src/devices/ILLMMsgs/idl_generated_code/yarp/dev/llm/ILLMMsgs.cpp
@@ -8,11 +8,13 @@
 // This is an automatically generated file.
 // It could get re-generated if the ALLOW_IDL_GENERATION flag is on.
 
-#include <ILLMMsgs.h>
+#include <yarp/dev/llm/ILLMMsgs.h>
 
 #include <yarp/os/idl/WireTypes.h>
 
 #include <algorithm>
+
+namespace yarp::dev::llm {
 
 // setPrompt helper class declaration
 class ILLMMsgs_setPrompt_helper :
@@ -1255,3 +1257,5 @@ bool ILLMMsgs::read(yarp::os::ConnectionReader& connection)
     }
     return false;
 }
+
+} // namespace yarp::dev::llm
