+++ src/devices/ILLMMsgs/idl_generated_code/yarp/dev/llm/return_getConversation.h
@@ -13,7 +13,9 @@
 
 #include <yarp/os/Wire.h>
 #include <yarp/os/idl/WireTypes.h>
-#include <Message.h>
+#include <yarp/dev/llm/Message.h>
+
+namespace yarp::dev::llm {
 
 class return_getConversation :
         public yarp::os::idl::WirePortable
@@ -61,5 +63,7 @@ private:
     bool nested_read_conversation(yarp::os::idl::WireReader& reader);
     bool nested_write_conversation(const yarp::os::idl::WireWriter& writer) const;
 };
+
+} // namespace yarp::dev::llm
 
 #endif // YARP_THRIFT_GENERATOR_STRUCT_RETURN_GETCONVERSATION_H
