+++ src/devices/LLM_nws_yarp/ILLMServerImpl.h
@@ -3,10 +3,10 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-#include <ILLMMsgs.h>
+#include <yarp/dev/llm/ILLMMsgs.h>
 #include <yarp/dev/ILLM.h>
 
-class ILLMRPCd : public ILLMMsgs
+class ILLMRPCd : public yarp::dev::llm::ILLMMsgs
 {
 private:
     yarp::dev::ILLM *m_iLlm = nullptr;
@@ -15,8 +15,8 @@ public:
     void setInterface(yarp::dev::ILLM *_iLlm) { m_iLlm = _iLlm; }
     // From IGPTMsgs
     bool setPrompt(const std::string &prompt) override;
-    return_readPrompt readPrompt() override;
-    return_ask ask(const std::string &question) override;
-    return_getConversation getConversation() override;
+    yarp::dev::llm::return_readPrompt readPrompt() override;
+    yarp::dev::llm::return_ask ask(const std::string &question) override;
+    yarp::dev::llm::return_getConversation getConversation() override;
     bool deleteConversation() override;
 };
